Store timestamps as integers of microseconds
Prevents precision loss when storing them in db.

+++ src/usdb_syncer/db/sql/setup_script.sql
@@ -3,7 +3,7 @@ BEGIN;
 CREATE TABLE meta (
     id INTEGER NOT NULL,
     version INTEGER NOT NULL,
-    ctime REAL NOT NULL,
+    ctime INTEGER NOT NULL,
     PRIMARY KEY (id)
 );
 
@@ -23,7 +23,7 @@ CREATE TABLE sync_meta (
     sync_meta_id INTEGER NOT NULL,
     song_id INTEGER NOT NULL,
     path TEXT NOT NULL,
-    mtime REAL NOT NULL,
+    mtime INTEGER NOT NULL,
     meta_tags TEXT NOT NULL,
     pinned BOOLEAN NOT NULL,
     PRIMARY KEY (sync_meta_id),
@@ -35,7 +35,7 @@ CREATE TABLE resource_file (
     sync_meta_id INTEGER NOT NULL,
     kind TEXT NOT NULL,
     fname TEXT NOT NULL,
-    mtime REAL NOT NULL,
+    mtime INTEGER NOT NULL,
     resource TEXT NOT NULL,
     PRIMARY KEY (sync_meta_id, kind),
     FOREIGN KEY(sync_meta_id) REFERENCES sync_meta (sync_meta_id) ON DELETE CASCADE
